% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reader.R
\name{csv_reader}
\alias{csv_reader}
\title{CSV data import}
\usage{
csv_reader(filepath="path", header, sep, dec, variable.characters, variable.integer, variable.date, date_format, timezone)
}
\arguments{
\item{filepath}{string Path of file}

\item{header}{logical csv file contain header TRUE else False. default TRUE}

\item{sep}{string. type of separator, by default ";"}

\item{dec}{string type of decimal by default "."}

\item{variable.characters}{vector of string. List of variables names corresponding to character
by default c("Genotype", "modality", "type_of_crown", "stage", "complete_module")}

\item{variable.integer}{vector of string List of variables names corresponding to integer
by default c("plant","order","no_visible_leaves","no_foliar_primordia","no_total_leaves","no_open_flowers",
"no_aborted_flowers","no_total_flowers","no_fruits","no_stolons","no_vegetative_bud","no_initiated_bud",
"no_floral_bud"),}

\item{variable.date}{vector of string List of variables names corresponding to date, by default date}

\item{date_format}{string format of the date, by default "\%d/\%m/\%Y"}

\item{timezone}{string. timezone of the experiment, by default "UTC"}
}
\value{
a Dataframe
}
\description{
This function allows to import csv data issue from data extraction at different scale from Web openalea.strawberry application
}
\examples{
csv_reader(filepath = "D:/Mes Donnees/BreedingValue/R/RStrawberryAnalysis/data/raw/Module_scale_data.csv",
                     header = TRUE,
                     sep = ";",
                     dec=".",
                     variable.characters = c("Genotype", "modality", "type_of_crown", "stage", "complete_module"), # nolint
                     variable.integer = c("plant",
                                       "order",
                                       "no_visible_leaves",
                                       "no_foliar_primordia",
                                       "no_total_leaves",
                                       "no_open_flowers",
                                       "no_aborted_flowers",
                                       "no_total_flowers",
                                       "no_fruits",
                                       "no_stolons",
                                       "no_vegetative_bud",
                                       "no_initiated_bud",
                                       "no_floral_bud"),
                 variable.date = c("date"),
                 date_format = "\%d/\%m/\%Y",
                 timezone= "UTC")
}
