% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{distribution.crown_type}
\alias{distribution.crown_type}
\title{Table of distribution of the type of crown}
\usage{
distribution.crown_type(data)
}
\arguments{
\item{data}{Dataframe. Data at module scale}
}
\value{
Dataframe. data containing for each crown type frequency (n) and relative frequency (prob) for each genotype and order
}
\description{
Contingency table of Type of Crowns distribution.
Return the contengency table in frequency or probability according frequency_type option
}
\examples{
    load("data/module_data.rdata")
    distribution.crown_type(data = module_data)
}
