% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{distribution.stage}
\alias{distribution.stage}
\title{Table of stage distribution at plant scale}
\usage{
distribution.stage(data)
}
\arguments{
\item{data}{Dataframe. Dataframe at module scale}
}
\value{
Dataframe. data containing stage frequency (n) and relative frequency (prob) of stage for each genotype and date
}
\description{
Table of distribution stage at plant scale from module scale data
}
\examples{
    load("data/module_data.rdata")
    distribution.sate(data = module_data)
}
