% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_axillary_production}
\alias{plot_axillary_production}
\title{Plot of axillary production accoring to node rank}
\usage{
plot_axillary_production(data, genotype, order = 0, variable = "n")
}
\arguments{
\item{data}{DataFrame. DataFrame issue from distribution.axillary_production function}

\item{genotype}{string. Genotype name selected}

\item{order}{int. Order selected, by default 0}

\item{variable}{string. "n" for frequency or "prob" for relative frequency}
}
\value{
plot with distribution of axillary production as function of node rank
}
\description{
Plot of axillary production according to genotype and order as function of node rank
}
\examples{
    load("data/node_data.rdata")
    dist <- distribution.axillary_production(node_data)
    plot_axillary_production(data = table, genotype = "Belle_et_Bonne", order = 0, variable = "n")
}
