% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_distribution.crown_type}
\alias{plot_distribution.crown_type}
\title{Crown type plot distribution}
\usage{
plot_distribution.crown_type(data, crown_type,variable="freq")
}
\arguments{
\item{data}{DataFrame. Crown type distribution data from distribution.crown_type function}

\item{crown_type}{string. type of crown ("Branch", "Extension")}

\item{variable}{string. "freq" for frequency or "rfreq" for relative frequency, by default "freq"}
}
\value{
plot of crown_type frequency or relative frequency distribution
}
\description{
Plot crown type distribution
}
\examples{
    load("data/module_data.rdata")
    dist <- distribution.crown_type(data = module_data)
    plot_distribution.crown_type(data=dist, crown_type = "Extension", variable = "freq")
}
