% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_pie.stage}
\alias{plot_pie.stage}
\title{Pie Plot distribution of stage at plant scale}
\usage{
plot_pie.stage(data, genotype, date, variable)
}
\arguments{
\item{data}{Dataframe. Dataframe issue from distribution stage function}

\item{genotype}{string. Genotype name selected}

\item{date}{string. Date selected}

\item{variable}{string. n for frequency and percent for percentage, by default n}
}
\value{
Pie plot
}
\description{
Pie plot distribution of stage at plant scale
}
\examples{
    load("data/module_data.rdata")
    df=distribution.stage(module_data)
    plot_pie.stage(data = df, genotype = "Belrubi", date = "2022-04-25",variable = "percent")

}
