% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_pointwisemean}
\alias{plot_pointwisemean}
\title{Plot pointwise mean}
\usage{
plot_pointwisemean(data, scale, variable, display_sd = TRUE)
}
\arguments{
\item{data}{Dataframe. Data from extraction at plant and module scale}

\item{scale}{string. "plant" or "module" according to data extraction}

\item{variable}{string. Variable plotted}

\item{display_sd}{logical. If you want display standard deviation or not, by default TRUE}
}
\value{
Dataframe
}
\description{
Plot pointwise mean
}
\examples{
    load("data/module_data.rdata")
    variables <- c("no_visible_leaves",
                         "no_foliar_primordia",
                         "no_total_leaves",
                         "no_open_flowers",
                         "no_aborted_flowers",
                         "no_total_flowers",
                         "no_fruits",
                         "no_stolons",
                         "no_vegetative_bud",
                         "no_initiated_bud",
                         "no_floral_bud")
    summarize_data <- summarize.extract_module(data = module_data, variables = variables)
    plot_pointwisemean(summarize_data, variable = "no_total_leaves",scale = "module", display_sd = FALSE)
}
