% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{summarize_extract_module}
\alias{summarize_extract_module}
\title{Table of mean and standard deviation according to genotype, date, modality, order}
\usage{
summarize_extract_module(data, variables)
}
\arguments{
\item{data}{Dataframe. Dataframe issue from extraction at module scale}

\item{variables}{list of string. list of character/string. Variables which mean and standard deviation will be calculated.}
}
\value{
Dataframe. where mean and standard deviation are compute for all variables
}
\description{
Calculate mean and standard variation according to Genotype, date, modality, order
}
\examples{
    load("data/module_data.rdata")
    variables <- c("no_visible_leaves",
                         "no_foliar_primordia",
                         "no_total_leaves",
                         "no_open_flowers",
                         "no_aborted_flowers",
                         "no_total_flowers",
                         "no_fruits",
                         "no_stolons",
                         "no_vegetative_bud",
                         "no_initiated_bud",
                         "no_floral_bud")
   summarize_data <- summarize_extract_module(data = module_data, variables = variables)

}
