% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{summarize_extract_node}
\alias{summarize_extract_node}
\title{Table of mean and standard deviation according to genotype, date, modality, order}
\usage{
summarize_extract_node(data,variables)
}
\arguments{
\item{data}{Dataframe. Dataframe issue from extraction at node scale}

\item{variables}{list string. Variables which mean and standard deviation will be calculated}
}
\value{
Dataframe
}
\description{
Calculate mean and standard variation according to Genotype, date, modality and order
}
\examples{
     load("data/node_data.rdata")

     node_variables <- c("no_visible_leaves","no_foliar_primordia","no_total_leaves",
     "no_open_flowers","no_aborted_flowers","no_total_flowers","no_fruits",
     "no_stolons","no_vegetative_bud","no_initiated_bud","no_floral_bud")

     summarize_extract_node(data = node_data, variables = node_variables)

}
