% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{summarize_extract_plant}
\alias{summarize_extract_plant}
\title{Table of mean and standard deviation according to genotype, date, modality}
\usage{
summarize_extract_plant(data,variables)
}
\arguments{
\item{data}{Dataframe. Dataframe issue from extraction at plant scale}

\item{variables}{list of character/string. Variables which mean and standard deviation will be calculated}
}
\value{
Dataframe. where mean and standard deviation are compute for all variable
}
\description{
Calculate mean and standard variation according to Genotype, date, modality
}
\examples{
   load("data/plant_data.rdata")
   plant_variables <- c("no_visible_leaves","no_foliar_primodium","no_total_leaves",
                        "no_open_flowers", "no_aborted_flowers","no_closed_flowers",
                        "no_total_flowers","no_fruits","no_stolons","no_vegetative_bud",
                        "no_initiated_bud","no_floral_bud","no_inflorescences",
                        "no_branch_crown","no_extension_crown","no_ramifications",
                        "order_max")

    summarize_extract_plant(data = plant_data, variables=plant_variables)

}
