% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{test.chi}
\alias{test.chi}
\title{Chi-test pairwise comparison to analyse crowns distribution by order}
\usage{
test.chi(data = data, order = 1, crown_type = "Branch")
}
\arguments{
\item{data}{Dataframe. Data issue from distribution.crown_type}

\item{order}{nt. Order for which analysis is performed.}

\item{crown_type}{string. "Branch" or "Extension" Crowns}
}
\value{
Dataframe containing pairwise chi-test output
}
\description{
@description reshape distribution data and perform pairwise chi-test comparison between genotype for one order from rstatix library
}
\examples{
    load("data/module_data.rdata")
    dist <- distribution.crown_type(data = module_data)
    test.chi(data = dist, order = 1, crown_type = "Branch")
}
