% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{test.kruskal}
\alias{test.kruskal}
\title{ANOVA on the Rank test (Kruskal Wallis)}
\usage{
test.kruskal(data, variable, trt, letter = TRUE, p.adj = "holm", alpha = 0.05, display = TRUE)
}
\arguments{
\item{data}{Dataframe. Dataframe from module scale extraction}

\item{variable}{string. response, name of the variable wish you execute comparison test}

\item{trt}{treatment Variables or intercept function}

\item{letter}{logical. TRUE display group with letters else FALSE, by default TRUE}

\item{p.adj}{Method for adjusting p values, by default "holm" ("none","holm","hommel", "hochberg", "bonferroni", "BH", "BY", "fdr")}

\item{alpha}{float. level signification, by default 0.05}

\item{display}{bool. print output, by default TRUE}
}
\value{
dataframe containing kruskal Wallis output
}
\description{
Kruskal Wallis Test from agricolae package.
}
\examples{
    load("data/module_data.rdata")
    test.kruskal(data = module_data,
                 variable = "no_total_leaves",
                 letter = FALSE,
                 trt = interaction(module_data[, "Genotype"], module_data[, "order"])
                 )

}
