% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{theme_publish}
\alias{theme_publish}
\title{Graphical Theme for publication}
\usage{
theme_publish(base_size = 12, base_family = "", base_line_size = 0.8)
}
\arguments{
\item{base_size}{int. text size}

\item{base_family}{sting. Police}

\item{base_line_size}{int. line size}
}
\value{
ggplot theme for publication
}
\description{
Graphical Theme for publication
}
